7/2/2019

+++ TestProject.Droid/Services/AudioService.cs
@@ -133,23 +133,23 @@ namespace TestProject.Droid.Services
 
         public void RenameFile(int id)
         {
-            path = Path.Combine(System.Environment.
-                GetFolderPath(System.Environment.
-                SpecialFolder.Personal), id.ToString() + TwitterUserId.Id_User + ".3gpp");
-
-            if (File.Exists(path))
+            var path = Path.Combine(System.Environment.
+                 GetFolderPath(System.Environment.
+                 SpecialFolder.Personal), id.ToString() + TwitterUserId.Id_User + ".3gpp");
+            if (File.Exists(_initialpath))
             {
-                File.Delete(path);
-                File.Move(_initialpath, path);
+                if (File.Exists(path))
+                {
+                    File.Delete(path);
+                    File.Move(_initialpath, path);
+                    File.Delete(_initialpath);
+                }
+                else
+                {
+                    File.Move(_initialpath, path);
+                    File.Delete(_initialpath);
+                }
             }
-            else
-            {
-                File.Move(_initialpath, path);
-            }
-            
-
-            File.Delete(_initialpath);
-
         }
 
         public void DeleteNullFile()
